+++ src/lightly_train/_task_models/dinov2_eomt_instance_segmentation/train_model.py
@@ -26,6 +26,7 @@ from lightly_train._data.instance_segmentation_dataset import (
     YOLOInstanceSegmentationDataArgs,
 )
 from lightly_train._data.task_data_args import TaskDataArgs
+from lightly_train._data.task_dataset import TaskDataset
 from lightly_train._metrics.instance_segmentation.task_metric import (
     InstanceSegmentationTaskMetric,
     InstanceSegmentationTaskMetricArgs,
@@ -109,6 +110,7 @@ class DINOv2EoMTInstanceSegmentationTrainArgs(TrainModelArgs):
         model_name: str,
         model_init_args: dict[str, Any],
         data_args: TaskDataArgs,
+        train_dataset: TaskDataset | None = None,
     ) -> None:
         if self.num_queries == "auto":
             num_queries = model_init_args.get("num_queries", 200)