+++ src/lightly_train/_task_models/dinov3_eomt_semantic_segmentation/train_model.py
@@ -26,6 +26,7 @@ from lightly_train._data.mask_semantic_segmentation_dataset import (
     MaskSemanticSegmentationDataArgs,
 )
 from lightly_train._data.task_data_args import TaskDataArgs
+from lightly_train._data.task_dataset import TaskDataset
 from lightly_train._metrics.semantic_segmentation.task_metric import (
     SemanticSegmentationTaskMetric,
     SemanticSegmentationTaskMetricArgs,
@@ -112,6 +113,7 @@ class DINOv3EoMTSemanticSegmentationTrainArgs(TrainModelArgs):
         model_name: str,
         model_init_args: dict[str, Any],
         data_args: TaskDataArgs,
+        train_dataset: TaskDataset | None = None,
     ) -> None:
         if self.num_queries == "auto":
             num_queries = model_init_args.get("num_queries", 100)