+++ src/lightly_train/_task_models/linear_semantic_segmentation/train_model.py
@@ -24,6 +24,7 @@ from lightly_train._data.mask_semantic_segmentation_dataset import (
     MaskSemanticSegmentationDataArgs,
 )
 from lightly_train._data.task_data_args import TaskDataArgs
+from lightly_train._data.task_dataset import TaskDataset
 from lightly_train._metrics.semantic_segmentation.task_metric import (
     SemanticSegmentationTaskMetric,
     SemanticSegmentationTaskMetricArgs,
@@ -74,6 +75,7 @@ class LinearSemanticSegmentationTrainArgs(TrainModelArgs):
         model_name: str,
         model_init_args: dict[str, Any],
         data_args: TaskDataArgs,
+        train_dataset: TaskDataset | None = None,
     ) -> None:
         if self.drop_path_rate == "auto":
             backbone_args = model_init_args.get("backbone_args", {})